improve appbar and drawer,clean code.
