+++ lib/view/screens/home/home.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:quick_routine/view/drawer/drawer.dart';
-import 'package:quick_routine/view/screens/config/config.dart';
+import 'package:quick_routine/view/common/appbar.dart';
+import 'package:quick_routine/view/common/drawer.dart';
 import 'package:quick_routine/view/screens/home/widget/task.dart';
 
 class Home extends StatefulWidget {
@@ -14,38 +14,7 @@ class _HomeState extends State<Home> {
     return Scaffold(
         backgroundColor: Colors.black54,
         drawer: buildDrawer(context),
-        appBar: PreferredSize(
-            preferredSize: Size.fromHeight(70.0),
-            child: AppBar(
-              //  iconTheme: new IconThemeData(size: 15.0,color: Colors.black54),
-              //         leading: IconButton(
-              //           icon: Icon(Icons.dehaze),
-              //           color: Colors.black54,
-              //           iconSize: 40.0,
-              //             onPressed: () {
-              //  //     Scaffold.of(context).openDrawer();
-
-              //     },
-
-              leading: Builder(builder: (BuildContext context) {
-                return IconButton(
-                    icon: Icon(Icons.dehaze),
-                    color: Colors.black54,
-                    iconSize: 40.0,
-                    onPressed: () {
-                      Scaffold.of(context).openDrawer();
-                    });
-              }),
-
-              title: Row(
-                verticalDirection: VerticalDirection.down,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Image.asset("assets/logo/logo-horizontal.png")
-                ],
-              ),
-              backgroundColor: Color.fromRGBO(192, 192, 192, 1.0),
-            )),
+        appBar: buildAppBar(context),
         body: Container(
           alignment: Alignment.bottomCenter,
           color: Colors.black12,
